StatSystemBaseImplementation
Add the base implementation of the stat system and now trying to include in constructors of state system so can be referenced across all states. Also, tried have a private variable to CharacterStat and then a getter and setter...seemed to cause issue so just going with a public variable for now.